rename some things and scaffold initial routing
