+++ packages/agents-ui/index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/client.tsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
