+++ packages/agents-ui/package.json
@@ -27,6 +27,8 @@
     "@radix-ui/react-tabs": "^1.1.1",
     "@radix-ui/react-tooltip": "^1.1.4",
     "@tanstack/react-query": "^5.67.3",
+    "@tanstack/react-router": "^1.95.5",
+    "@tanstack/react-router-devtools": "^1.114.29",
     "@uiw/codemirror-theme-basic": "^4.23.7",
     "@uiw/codemirror-theme-duotone": "^4.23.6",
     "@uiw/react-codemirror": "^4.23.6",
@@ -43,19 +45,20 @@
     "@biomejs/biome": "^1.9.4",
     "@shadcn/ui": "^0.0.4",
     "@tailwindcss/container-queries": "^0.1.1",
-    "@types/react-dom": "^19.0.4",
+    "@tanstack/router-plugin": "^1.95.5",
     "@types/react": "^19.0.10",
-    "@vitejs/plugin-react-swc": "^3.8.0",
+    "@types/react-dom": "^19.0.4",
     "@vitejs/plugin-react": "^4.3.4",
+    "@vitejs/plugin-react-swc": "^3.8.0",
     "autoprefixer": "^10.4.21",
     "dotenv": "^16.4.7",
     "postcss": "^8.5.3",
     "tailwind-merge": "^2.5.5",
-    "tailwindcss-animate": "^1.0.7",
     "tailwindcss": "^3.4.15",
+    "tailwindcss-animate": "^1.0.7",
     "typescript": "^5.8.2",
+    "vite": "^6.2.1",
     "vite-plugin-node-polyfills": "^0.23.0",
-    "vite-plugin-svgr": "^4.3.0",
-    "vite": "^6.2.1"
+    "vite-plugin-svgr": "^4.3.0"
   }
 }
