+++ packages/agents-ui/src/components/index.ts
@@ -1,2 +1,1 @@
-export { App } from "./App";
 export { ThemeProvider } from "./ThemeProvider";
